+++ Locale/esES.lua
@@ -116,7 +116,7 @@ CLM.L["Akil'zon"] = "Akil'zon"
 CLM.L["Al'ar"] = "Al'ar"
 -- Modules/AutoAwardManager/EncounterIDs.lua:125
 CLM.L["Algalon the Observer"] = "Algalon el observador"
--- Modules/BiddingManager/GUI.lua:219
+-- Modules/BiddingManager/GUI.lua:221
 CLM.L["All In"] = "Todo incluido"
 -- Migration.lua:28
 CLM.L["All migration entries were commited and executed. Congratulations!"] = "Todas las entradas de migración se confirmaron y ejecutaron. ¡Felicidades!"
@@ -336,9 +336,9 @@ CLM.L["Bid %s denied!"] = "¡Oferta %s rechazada!"
 CLM.L["Bid cancelling not allowed"] = "Cancelación de oferta no permitida"
 -- Modules/AuctionManager/AuctionManager.lua:676
 CLM.L["Bid increment too low"] = "Incremento de oferta demasiado bajo"
--- Modules/BiddingManager/GUI.lua:159
+-- Modules/BiddingManager/GUI.lua:157
 CLM.L["Bid input values as Main spec bid."] = "Valores de entrada de oferta como oferta de especificaciones principales."
--- Modules/BiddingManager/GUI.lua:170
+-- Modules/BiddingManager/GUI.lua:168
 CLM.L["Bid input values as Off spec bid."] = "Valores de entrada de oferta como oferta de especificación OFF."
 -- Modules/AuctionManager/AuctionManager.lua:244
 CLM.L["Bid tiers: %s."] = "Niveles de ofertas: %s."
@@ -346,9 +346,9 @@ CLM.L["Bid tiers: %s."] = "Niveles de ofertas: %s."
 CLM.L["Bid too high"] = "Oferta demasiado alta"
 -- Modules/AuctionManager/AuctionManager.lua:673
 CLM.L["Bid too low"] = "Oferta demasiado baja"
--- Modules/BiddingManager/GUI.lua:149
+-- Modules/BiddingManager/GUI.lua:147
 CLM.L["Bid value"] = "Valor de la oferta"
--- Modules/BiddingManager/GUI.lua:220
+-- Modules/BiddingManager/GUI.lua:222
 CLM.L["Bid your current DKP (%s)."] = "Oferte su DKP actual (%s)."
 -- Modules/AuctionManager/GUI.lua:221
 CLM.L["Bid"] = "Oferta de compra"
@@ -358,8 +358,10 @@ CLM.L["Bidding over current standings not allowed"] = "No se permite pujar sobre
 CLM.L["Bidding while below minimum standings not allowed"] = "Licitación de licitación por debajo de la clasificación mínima no permitida"
 -- MinimapIcon.lua:138
 -- Modules/BiddingManager/BiddingManager.lua:47
--- Modules/BiddingManager/GUI.lua:397
--- Modules/BiddingManager/GUI.lua:558
+-- Modules/BiddingManager/GUI.lua:413
+-- Modules/BiddingManager/GUI.lua:506
+-- Modules/BiddingManager/GUI.lua:514
+-- Modules/BiddingManager/GUI.lua:576
 -- Modules/RosterManager/Options.lua:874
 CLM.L["Bidding"] = "Ofertas"
 -- Modules/AuctionHistoryManager/GUI.lua:137
@@ -398,11 +400,11 @@ CLM.L["Buru the Gorger"] = "Buru el Manducador"
 CLM.L["Button Names"] = "Nombres de botones"
 -- Modules/AutoAwardManager/EncounterIDs.lua:31
 CLM.L["C'Thun"] = "C'Thun"
--- Modules/BiddingManager/GUI.lua:201
--- Modules/BiddingManager/GUI.lua:352
+-- Modules/BiddingManager/GUI.lua:199
+-- Modules/BiddingManager/GUI.lua:365
 CLM.L["Cancel your bid."] = "Cancele su oferta."
--- Modules/BiddingManager/GUI.lua:200
--- Modules/BiddingManager/GUI.lua:351
+-- Modules/BiddingManager/GUI.lua:198
+-- Modules/BiddingManager/GUI.lua:364
 CLM.L["Cancel"] = "Cancelar"
 -- Modules/RosterManager/Options.lua:617
 CLM.L["Change roster name."] = "Cambiar el nombre del Roster."
@@ -438,7 +440,7 @@ CLM.L["Classic"] = "Classic"
 CLM.L["Clear all classes."] = "Borrar todas las clases."
 -- Integrations/GUI.lua:470
 CLM.L["Clear output"] = "Borrar salida"
--- Modules/BiddingManager/GUI.lua:86
+-- Modules/BiddingManager/GUI.lua:84
 CLM.L["Close on bid"] = "Cerrar en la oferta"
 -- Utils.lua:755
 CLM.L["Common"] = "Común"
@@ -477,6 +479,7 @@ CLM.L["Creates new roster with default configuration"] = "Crea un nuevo roster c
 CLM.L["Ctrl + Alt"] = "Ctrl + Alt"
 -- Global/GlobalConfigs.lua:365
 CLM.L["Ctrl"] = "Ctrl"
+-- Modules/BiddingManager/GUI.lua:514
 -- Modules/AuctionManager/GUI.lua:225
 CLM.L["Current"] = "Actual"
 -- Alerts/Alerts.lua:17
@@ -870,8 +873,8 @@ CLM.L["Item value mode"] = "Modo de valor del artículo"
 CLM.L["Item value must be positive"] = "El valor del artículo debe ser positivo"
 -- Modules/RosterManager/Options.lua:927
 CLM.L["Item value overrides"] = "Reemplazos del valor del artículo"
--- Modules/BiddingManager/GUI.lua:140
--- Modules/BiddingManager/GUI.lua:280
+-- Modules/BiddingManager/GUI.lua:138
+-- Modules/BiddingManager/GUI.lua:292
 -- Modules/AuctionManager/GUI.lua:308
 CLM.L["Item"] = "Articulo"
 -- Modules/AutoAwardManager/EncounterIDs.lua:104
@@ -970,7 +973,7 @@ CLM.L["Lucifron"] = "Lucifron"
 CLM.L["M'uru"] = "M'uru"
 -- Integrations/Integrations.lua:133
 -- Integrations/Integrations.lua:138
--- Modules/BiddingManager/GUI.lua:158
+-- Modules/BiddingManager/GUI.lua:156
 CLM.L["MS"] = "MS"
 -- Modules/AutoAwardManager/EncounterIDs.lua:34
 CLM.L["Maexxna"] = "Maexxna"
@@ -1169,11 +1172,11 @@ CLM.L["Note to be added to award. Max 25 characters. It is recommended to not in
 CLM.L["Note"] = "Nota"
 -- Modules/AutoAwardManager/EncounterIDs.lua:39
 CLM.L["Noth the Plaguebringer"] = "Noth el Pesteador"
--- Modules/BiddingManager/GUI.lua:185
--- Modules/BiddingManager/GUI.lua:336
-CLM.L["Notify that you are passing on the item. Cancels any existing bids."] = "Notifique que está pasando el artículo. Cancela cualquier oferta existente."
 -- Modules/BiddingManager/GUI.lua:183
--- Modules/BiddingManager/GUI.lua:334
+-- Modules/BiddingManager/GUI.lua:349
+CLM.L["Notify that you are passing on the item. Cancels any existing bids."] = "Notifique que está pasando el artículo. Cancela cualquier oferta existente."
+-- Modules/BiddingManager/GUI.lua:181
+-- Modules/BiddingManager/GUI.lua:347
 CLM.L["Notify that you are passing on the item."] = "Notifique que está pasando el artículo."
 -- Integrations/GUI.lua:149
 CLM.L["November"] = "Noviembre"
@@ -1181,7 +1184,7 @@ CLM.L["November"] = "Noviembre"
 CLM.L["Num"] = "Número"
 -- Integrations/Integrations.lua:143
 -- Integrations/Integrations.lua:148
--- Modules/BiddingManager/GUI.lua:169
+-- Modules/BiddingManager/GUI.lua:167
 CLM.L["OS"] = "OS"
 -- Integrations/GUI.lua:148
 CLM.L["October"] = "Octubre"
@@ -1240,8 +1243,8 @@ CLM.L["PR"] = "PR"
 CLM.L["Paladin"] = "Paladín"
 -- Modules/GUI/UnifiedGUI_Raids.lua:289
 CLM.L["Participated"] = "participó"
--- Modules/BiddingManager/GUI.lua:180
--- Modules/BiddingManager/GUI.lua:331
+-- Modules/BiddingManager/GUI.lua:178
+-- Modules/BiddingManager/GUI.lua:344
 CLM.L["Pass"] = "Pasar"
 -- Modules/AuctionManager/AuctionManager.lua:679
 CLM.L["Passing after bidding not allowed"] = "Pasar después de pujar no está permitido"
@@ -1664,11 +1667,11 @@ CLM.L["Time"] = "Hora"
 CLM.L["Timetravel"] = "Viaje en el tiempo"
 -- Modules/AuctionHistoryManager/GUI.lua:275
 CLM.L["Toggle Auction History window display"] = "Alternar la visualización de la ventana Historial de subastas"
--- Modules/AuctionManager/GUI.lua:713
+-- Modules/AuctionManager/GUI.lua:719
 CLM.L["Toggle Auctioning window display"] = "Alternar la visualización de la ventana de Subastas"
 -- Modules/BiddingManager/BiddingManager.lua:51
 CLM.L["Toggle Bidding auto-open"] = "Alternar ofertas de apertura automática"
--- Modules/BiddingManager/GUI.lua:559
+-- Modules/BiddingManager/GUI.lua:577
 CLM.L["Toggle Bidding window display"] = "Alternar la visualización de la ventana de ofertas"
 -- Modules/LootQueueManager/GUI.lua:263
 CLM.L["Toggle Loot Queue window display"] = "Alternar la visualización de la ventana Loot Queue"
@@ -1681,7 +1684,7 @@ CLM.L["Toggle auto open and auto close on auction start and stop"] = "Alternar e
 CLM.L["Toggle changelog window display"] = "Alternar la visualización de la ventana de registro de cambios"
 -- Modules/Changelog/GUI.lua:47
 CLM.L["Toggle changelog"] = "Alternar registro de cambios"
--- Modules/BiddingManager/GUI.lua:87
+-- Modules/BiddingManager/GUI.lua:85
 CLM.L["Toggle closing bidding UI after submitting bid."] = "Alternar la interfaz de usuario de la oferta de cierre después de presentar la oferta."
 -- Integrations/GUI.lua:509
 CLM.L["Toggle export window display"] = "Alternar visualización de la ventana de exportación"
@@ -1770,7 +1773,7 @@ CLM.L["Val'kyr Twins"] = "Val'kyr gemelos"
 CLM.L["Valithria Dreamwalker"] = "Valithria Dreamwalker"
 -- Modules/AuctionManager/GUI.lua:359
 CLM.L["Value ranges"] = "Rangos de valores"
--- Modules/BiddingManager/GUI.lua:150
+-- Modules/BiddingManager/GUI.lua:148
 CLM.L["Value you want to bid. Press Enter or click Okay button to accept."] = "Valor que desea ofertar. Presione Entrar o haga clic en el botón OK para aceptar. ."
 -- Modules/GUI/UnifiedGUI_History.lua:146
 CLM.L["Value"] = "Valor"